few fixes to run trainning

+++ MPNet/AE/data_loader.py
@@ -14,7 +14,7 @@ def load_dataset(N=30000,NP=1800):
 	obstacles=np.zeros((N,2800),dtype=np.float32)
 	for i in range(0,N):
 		temp=np.fromfile('../dataset2/obs_cloud/obc'+str(i)+'.dat')
-		temp=temp.reshape(len(temp)/2,2)
+		temp=temp.reshape(int(len(temp)/2),2)
 		obstacles[i]=temp.flatten()
